operator: Hidden flags for egress gateway and LRPs
[ upstream commit 80f61ad1b4236c287f274e687c9e96a8afdc9da6 ]

Some of the CRD registrations (e.g., CENP and CLRP) are conditioned on
feature flags. In some cases, that registration happens in the operator,
in which case it needs to be possible to enable the features there.

Reproduced and tested in the development VM where the operator is in
charge of registering the CRDs. Without this patch, the CRDs are never
created and the agent waits indefinitely when e.g. the egress gateway is
enabled. With this patch, the egress gateway can be enabled in the
operator via a flag to trigger the registration of the CRD.

Suggested-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Timo Beckers <timo@isovalent.com>