egressgateway: Allow several CENPs with same egress IP
[ upstream commit b717428f8895260e337416eec39c52c5d861b7cc ]

In Commit a783f7c5 ("k8s/cilium Event handlers and processing logic for
egress nat policy"), having several CENPs with the same egress IP was
explicitly disallowed. As explained in [1], the rationale was that those
egress IPs are sometimes associated with security identities in external
applications. However, this limitation is also preventing legitimate use
cases. Users who would like to prevent this can implement it on top of
Cilium.

This commit removes the restriction.

1 - https://github.com/cilium/cilium/issues/17678
Fixes: a783f7c5 ("k8s/cilium Event handlers and processing logic for egress nat policy")
Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Timo Beckers <timo@isovalent.com>